requirements.txt
python-telegram-bot
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0
yt-dlp==2023.12.30
asyncio
logging
